+++ server/data/sampleCommits.ts
@@ -1,4 +1,4 @@
-import { CommitSummary } from "../types/CommitSummary"
+import { CommitSummary } from "../types/CommitSummary";
 
 export const sampleCommits: CommitSummary[] = [
   {
@@ -7,6 +7,7 @@ export const sampleCommits: CommitSummary[] = [
     message: "Merge branch 'dc/wip'",
     linesAdded: 102,
     linesRemoved: 43,
+    repo: "fractal-bootcamp/hackathon-kitchen-gossip",
     filesChangedNum: 5,
     filesChangedNames:
       "server/cli.ts, server/justfile, server/routes/status.ts, server/services/callGithub.ts, server/services/reviewer.ts",
@@ -17,6 +18,7 @@ export const sampleCommits: CommitSummary[] = [
     message: "change error handling",
     linesAdded: 52,
     linesRemoved: 10,
+    repo: "fractal-bootcamp/hackathon-kitchen-gossip",
     filesChangedNum: 5,
     filesChangedNames:
       "server/cli.ts, server/justfile, server/routes/status.ts, server/services/callGithub.ts, server/services/reviewer.ts",
@@ -27,6 +29,7 @@ export const sampleCommits: CommitSummary[] = [
     message: "adding cli",
     linesAdded: 54,
     linesRemoved: 37,
+    repo: "fractal-bootcamp/hackathon-kitchen-gossip",
     filesChangedNum: 3,
     filesChangedNames:
       "server/cli.ts, server/justfile, server/services/callGithub.ts",
@@ -37,229 +40,9 @@ export const sampleCommits: CommitSummary[] = [
     message: "file move",
     linesAdded: 4,
     linesRemoved: 4,
+    repo: "fractal-bootcamp/hackathon-kitchen-gossip",
     filesChangedNum: 2,
     filesChangedNames:
       "server/services/callGithub.ts, server/services/getRepos.ts",
   },
-  {
-    user: "yab",
-    time: "",
-    message: "fixed paths",
-    linesAdded: 2,
-    linesRemoved: 2,
-    filesChangedNum: 1,
-    filesChangedNames: "server/services/callGithub.ts",
-  },
-  {
-    user: "dcsan",
-    time: "",
-    message: "move call github",
-    linesAdded: 0,
-    linesRemoved: 0,
-    filesChangedNum: 1,
-    filesChangedNames: "server/services/callGithub.ts",
-  },
-  {
-    user: "dcsan",
-    time: "",
-    message: "Merge branch 'main' into dc/wip",
-    linesAdded: 74,
-    linesRemoved: 18,
-    filesChangedNum: 3,
-    filesChangedNames:
-      "server/callGithub.ts, server/data/dummyData.ts, server/types/CommitSummary.ts",
-  },
-  {
-    user: "Lui",
-    time: "",
-    message:
-      "Merge pull request #2 from fractal-bootcamp/calling-github-p2\n\nCalling GitHub p2",
-    linesAdded: 117,
-    linesRemoved: 58,
-    filesChangedNum: 3,
-    filesChangedNames:
-      "server/callGithub.ts, server/data/dummyData.ts, server/types/CommitSummary.ts",
-  },
-  {
-    user: "dcsan",
-    time: "",
-    message: "wip",
-    linesAdded: 32,
-    linesRemoved: 16,
-    filesChangedNum: 6,
-    filesChangedNames:
-      "server/data/dummyData.ts, server/package-lock.json, server/package.json, server/services/github.ts, server/services/reviewer.ts, server/utils/getEnv.ts",
-  },
-  {
-    user: "yab",
-    time: "",
-    message: "updated type",
-    linesAdded: 8,
-    linesRemoved: 6,
-    filesChangedNum: 3,
-    filesChangedNames:
-      "server/callGithub.ts, server/data/dummyData.ts, server/types/CommitSummary.ts",
-  },
-  {
-    user: "yab",
-    time: "",
-    message: "full function working",
-    linesAdded: 47,
-    linesRemoved: 10,
-    filesChangedNum: 1,
-    filesChangedNames: "server/callGithub.ts",
-  },
-  {
-    user: "yab",
-    time: "",
-    message: "getRecentCommitList",
-    linesAdded: 21,
-    linesRemoved: 4,
-    filesChangedNum: 1,
-    filesChangedNames: "server/callGithub.ts",
-  },
-  {
-    user: "yab",
-    time: "",
-    message: "moved new data to dummyData",
-    linesAdded: 49,
-    linesRemoved: 46,
-    filesChangedNum: 3,
-    filesChangedNames:
-      "server/callGithub.ts, server/data/dummyData.ts, server/types/CommitSummary.ts",
-  },
-  {
-    user: "dcsan",
-    time: "",
-    message: "Merge branch 'main' into edge",
-    linesAdded: 576,
-    linesRemoved: 25,
-    filesChangedNum: 16,
-    filesChangedNames:
-      ".gitignore, docs/api-ex.json, server/.env.example, server/data/dummyData.ts, server/data/postExample.ts, server/data/prompts.ts, server/http/test.http, server/justfile, server/package-lock.json, server/package.json, server/routes/status.ts, server/server.ts, server/services/github.ts, server/services/reviewer.ts, server/services/slack.ts, server/types/shared.d.ts",
-  },
-  {
-    user: "yab",
-    time: "",
-    message: "added dummy data",
-    linesAdded: 26,
-    linesRemoved: 3,
-    filesChangedNum: 1,
-    filesChangedNames: "server/callGithub.ts",
-  },
-  {
-    user: "dcsan",
-    time: "",
-    message: "Merge branch 'dc/status-api'",
-    linesAdded: 576,
-    linesRemoved: 25,
-    filesChangedNum: 16,
-    filesChangedNames:
-      ".gitignore, docs/api-ex.json, server/.env.example, server/data/dummyData.ts, server/data/postExample.ts, server/data/prompts.ts, server/http/test.http, server/justfile, server/package-lock.json, server/package.json, server/routes/status.ts, server/server.ts, server/services/github.ts, server/services/reviewer.ts, server/services/slack.ts, server/types/shared.d.ts",
-  },
-  {
-    user: "dcsan",
-    time: "",
-    message: "atransform funcs",
-    linesAdded: 92,
-    linesRemoved: 6,
-    filesChangedNum: 5,
-    filesChangedNames:
-      "server/data/postExample.ts, server/routes/status.ts, server/services/reviewer.ts, server/services/slack.ts, server/types/shared.d.ts",
-  },
-  {
-    user: "yab",
-    time: "",
-    message: "full skeleton",
-    linesAdded: 62,
-    linesRemoved: 0,
-    filesChangedNum: 1,
-    filesChangedNames: "server/callGithub.ts",
-  },
-  {
-    user: "dcsan",
-    time: "",
-    message: ".env.example",
-    linesAdded: 2,
-    linesRemoved: 0,
-    filesChangedNum: 1,
-    filesChangedNames: "server/.env.example",
-  },
-  {
-    user: "dcsan",
-    time: "",
-    message: "first comparison",
-    linesAdded: 310,
-    linesRemoved: 38,
-    filesChangedNum: 10,
-    filesChangedNames:
-      ".gitignore, server/data/dummyData.ts, server/data/prompts.ts, server/http/test.http, server/package-lock.json, server/package.json, server/routes/status.ts, server/services/github.ts, server/services/reviewer.ts, server/types/shared.d.ts",
-  },
-  {
-    user: "yab",
-    time: "",
-    message: "adding sample data",
-    linesAdded: 116,
-    linesRemoved: 0,
-    filesChangedNum: 1,
-    filesChangedNames: "sampleData/sampleGHresponse",
-  },
-  {
-    user: "dcsan",
-    time: "",
-    message: "add basic status msg",
-    linesAdded: 44,
-    linesRemoved: 5,
-    filesChangedNum: 4,
-    filesChangedNames:
-      "docs/api-ex.json, server/data/dummyData.ts, server/routes/status.ts, server/services/github.ts",
-  },
-  {
-    user: "Lui",
-    time: "",
-    message:
-      "Merge pull request #1 from fractal-bootcamp/calling-github\n\nadding sample data",
-    linesAdded: 2,
-    linesRemoved: 0,
-    filesChangedNum: 1,
-    filesChangedNames: "server/server.ts",
-  },
-  {
-    user: "dcsan",
-    time: "",
-    message: "added first endpoint",
-    linesAdded: 172,
-    linesRemoved: 20,
-    filesChangedNum: 5,
-    filesChangedNames:
-      "server/justfile, server/package-lock.json, server/package.json, server/routes/status.ts, server/server.ts",
-  },
-  {
-    user: "hendeaux",
-    time: "",
-    message: "Update README.md",
-    linesAdded: 1,
-    linesRemoved: 0,
-    filesChangedNum: 1,
-    filesChangedNames: "README.md",
-  },
-  {
-    user: "yab",
-    time: "",
-    message: "simple merge",
-    linesAdded: 2,
-    linesRemoved: 0,
-    filesChangedNum: 1,
-    filesChangedNames: "server/server.ts",
-  },
-  {
-    user: "yab",
-    time: "",
-    message: "init",
-    linesAdded: 5147,
-    linesRemoved: 0,
-    filesChangedNum: 25,
-    filesChangedNames:
-      ".env, .env.test, .eslintrc.cjs, .gitignore, README.md, docker-compose.test.yml, docker-compose.yml, index.html, package-lock.json, package.json, prisma/schema.prisma, public/vite.svg, server/package-lock.json, server/package.json, server/server.ts, src/App.css, src/App.tsx, src/assets/react.svg, src/index.css, src/main.tsx, src/vite-env.d.ts, tsconfig.app.json, tsconfig.json, tsconfig.node.json, vite.config.ts",
-  },
-]
+];
