+++ server/routes/status.ts
@@ -1,18 +1,18 @@
 // status check
 
-import { getReviewStatus } from "../services/reviewer"
-import { postText } from "../services/slack"
-import { ReviewStatus } from "../types/shared"
+import { getReviewStatus } from "../services/reviewer";
+import { postText } from "../services/slack/postText";
+import { ReviewStatus } from "../types/shared";
 
 export function mount(app) {
   app.get("/api/status", async (req, res) => {
-    const reviewStatus: ReviewStatus = await getReviewStatus()
-    await postText(reviewStatus.reviews)
+    const reviewStatus: ReviewStatus = await getReviewStatus();
+    await postText(reviewStatus.reviews);
     // await postText(review.text)
-    console.log("STATUS called")
+    console.log("STATUS called");
     res.json({
       message: "status",
       reviewStatus,
-    })
-  })
+    });
+  });
 }
