--- server/services/github.ts
@@ -1,7 +1,0 @@
-import { CommitSummary } from "../types/CommitSummary"
-import { githubStatus, sampleCommitSummaries } from "../data/dummyData"
-import { GithubStatusList } from "../types/shared"
-
-export async function getStatus(opts?: any): Promise<CommitSummary[]> {
-  return sampleCommitSummaries
-}
