+++ server/services/slack/postText.ts
@@ -1,14 +1,14 @@
-import { getEnv } from "../utils/getEnv"
+import { getEnv } from "../../utils/getEnv";
 
 export async function postText(text: string) {
-  const url = getEnv("SLACK_WEBHOOK")
+  const url = getEnv("SLACK_WEBHOOK");
 
   const payload = {
     channel: "#kitchen-gossip",
     username: "superbot",
     icon_emoji: ":bot:",
     text: text,
-  }
+  };
 
   const response = await fetch(url, {
     method: "POST",
@@ -16,11 +16,11 @@ export async function postText(text: string) {
       "Content-Type": "application/json",
     },
     body: JSON.stringify(payload),
-  })
+  });
 
   if (!response.ok) {
-    throw new Error("Failed to post to Slack")
+    throw new Error("Failed to post to Slack");
   } else {
-    console.log("Posted to Slack", response.status)
+    console.log("Posted to Slack", response.status);
   }
 }
